Restructure and add example program
