+++ CMLParser.Example/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace CMLParser.Example
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
